Fix bug that leads to segmentation fault crash in KeytroLTestApp
How to reproduce:
 1. Select a computer key
 2. Press Enter or Return to start renaming
 3. Click on the selected computer key inside the 5 pixel margin
 around the TextInput to finish renaming
 4. Crash

On one hand, in step 3, the activeFocus was given back to the selected
computer key. On the other hand, when renaming a computer key, the
computer key with the old label is deleted and a new one with the
desired label is added (see KL::KeyboardLayoutViewModel::rename()).
These two events resulted in the crash since the activeFocus was
basically given to an Item that would be deleted in the next frame.

To fix it, only give the activeFocus to a computer key if it is not
already selected.