Fix console spamming when loading an XML file as first action
Line 181, mouseArea.selectedComputerKey is compared without coercion
against 'null'. If it is 'undefined', then the comparison will amount
to 'true' and the next line will try to access its modelIndex
property, leading to a TypeError warning printed on the console.
This situation only happens when loading an XML file right after
starting KeytroLTestApp, since updateSelectedComputerKey() has not
been called yet.

Set mouseArea.selectedComputerKey's default value to 'null' to ensure
that it is either 'null' or a ComputerKey.