Fix bug that prevents finishing renaming by pressing Enter or Return
How to reproduce:
 1. Select a computer key
 2. Press Enter or Return to start renaming
 3. Press Enter or Return to stop renaming
Expected: the TextInput cursor disappears
Actual: the TextInput cursor is still visible

In 8c9ae77 KL::KeyboardLayoutViewModel::rename() was changed to not
delete the selected key and create a new one. This means that in step
3 the key press triggers the TextInput.editingFinished() signal, but
then is processed by the Keys.onPressed handler line 184, since the
computer key still has the focus. So the TextInput ends up getting the
activeFocus back.

To fix it, only give the activeFocus to the TextInput if the computer
key has it, otherwise give it back to the computer key.