Only give activeFocus when using the mouse left button
That way, renaming a computer key is enabled by pressing on it with
the left button only. The right button will be used in a subsequent
commit to enable binding a virtual computer key to a physical one.

Adapt the condition to select the computer keys border color, since
selected and activeFocus are not equivalent anymore. Use activeFocus
to express that a computer key with a black border can be renamed by
pressing Enter or Return, or is currently being renamed.